rollback useGetTranslations
