+++ src/models/hooks/queries/index.ts
@@ -6,7 +6,6 @@ export * from "./useOffers";
 export * from "./useAdminTasks";
 export * from "./useCurrentUser";
 export * from "./useTranslations";
-export * from "./useGetTranslations";
 export * from "./useMyApplicantProfile";
 export * from "./useMyCompanyProfile";
 export * from "./useCareers";
